requirements.txt
fastapi
uvicorn[standard]
pycryptodome
python-multipart
